Refactor recovery policy handling and introduce new query for active alert group hashes
- Added a new `recoveryPolicyTypeSchema` to define recovery policy types in the rule data schema.
- Updated the `recoveryPolicySchema` to utilize the new schema for type validation.
- Moved the `getActiveAlertGroupHashesQuery` function to a new `queries.ts` file for better organization.
- Adjusted imports in `build_alert_events.ts` and `create_recovery_events_step.ts` to reflect the new file structure.
- Enhanced the `getActiveAlertGroupHashesQuery` to ensure it only retrieves group hashes with non-null episode statuses.
